+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - changed system prompt to XML structure for better instruction following across the various models
 - fixed issue with text jumping in Prompt field when typing fast
+- notifications correcly using selected theme
 
 ## [0.31.0]